+++ packaging/flatpak/fretman.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec /app/fretman/fretboard_trainer "$@"